W-14547524: Unable to read dataType variable (#13026)
* TD-0171856: Unable to read dataType variable
* Move DataType implementations to privileged api
* serialization backwards compatibility
* fix equals
