+++ core-tests/src/test/java/org/mule/runtime/core/internal/transformer/type/DataTypeBuilderTestCase.java
@@ -41,10 +41,10 @@ import org.mule.runtime.api.metadata.DataTypeParamsBuilder;
 import org.mule.runtime.api.metadata.FunctionDataType;
 import org.mule.runtime.api.metadata.FunctionParameter;
 import org.mule.runtime.api.metadata.TypedValue;
-import org.mule.runtime.core.internal.metadata.DefaultCollectionDataType;
-import org.mule.runtime.core.internal.metadata.DefaultFunctionDataType;
-import org.mule.runtime.core.internal.metadata.DefaultMapDataType;
-import org.mule.runtime.core.internal.metadata.SimpleDataType;
+import org.mule.runtime.core.privileged.metadata.DefaultCollectionDataType;
+import org.mule.runtime.core.privileged.metadata.DefaultFunctionDataType;
+import org.mule.runtime.core.privileged.metadata.DefaultMapDataType;
+import org.mule.runtime.core.privileged.metadata.SimpleDataType;
 import org.mule.tck.junit4.AbstractMuleTestCase;
 import org.mule.tck.probe.JUnitLambdaProbe;
 import org.mule.tck.probe.PollingProber;
