+++ tests/performance/src/main/java/org/mule/DataTypeBenchmark.java
@@ -9,7 +9,7 @@ package org.mule;
 import static java.util.concurrent.TimeUnit.NANOSECONDS;
 
 import org.mule.runtime.api.metadata.DataType;
-import org.mule.runtime.core.internal.metadata.DefaultDataTypeBuilder;
+import org.mule.runtime.core.privileged.metadata.DefaultDataTypeBuilder;
 
 import org.openjdk.jmh.annotations.Benchmark;
 import org.openjdk.jmh.annotations.OutputTimeUnit;
